Clean up basic agent call, update README
